add process_attestations to process_operations
